+++ src/console-proxy.ts
@@ -1,71 +1,92 @@
-import { App } from "obsidian";
-import { NoteLogger } from "./note-logger";
+import { LogEvent, UnhandledRejectionEvent } from "./types";
+import { prefixMsg } from "./utils";
 
-// Storing the original `console` object so we can restore it later
+// Storing the original `window.*` object so we can restore them later
 const WINDOW_CONSOLE = window.console;
 
-type ConsoleProxyOptions = {
-  app: App;
-  logger: NoteLogger;
-};
-
+/**
+ * A class that sets up a proxy for the `console` object to intercept log
+ * events, which are added to the passed-in queue.
+ */
 export class ConsoleProxy {
-  private readonly app: App;
-  private readonly logger: NoteLogger;
-
-  constructor(options: ConsoleProxyOptions) {
-    const { app, logger } = options;
-    this.app = app;
-    this.logger = logger;
-  }
+  constructor(
+    private queue: LogEvent[],
+  ) {}
 
   // Installing the console proxy object and the listener for uncaught errors
-  public setup(): void {
+  public setup(): ConsoleProxy {
     // Preventing scoping issues in the `Proxy` object
-    const { logger } = this;
+    const self = this;
+
+    const handler: ProxyHandler<Console> = {
+      get(target, prop) {
+        // Intercept the property access
+        const property = (<any> target)[prop];
+
+        if (typeof property === "function") {
+          // Wrap the original method with the extra logging behavior
+          return (...args: any[]) => {
+            // Get the sender of the log event by parsing the stack trace
+            const sender = new Error().stack
+              ?.split("\n").at(2)
+              ?.replace(/^.+\((.+?)\).*$/, "$1")
+              .replace("app://obsidian.md/", "")
+              .trim();
+
+            self.storeEvent(prop.toString(), sender, args);
+
+            // Forward the method call to the original `console` method
+            return property.apply(target, args);
+          };
+        }
+
+        // Return non-function properties as is
+        return property;
+      },
+    };
 
     // Set up a new `Proxy` object to intercept any `console` calls
-    const consoleProxy = new Proxy(
-      window.console,
-      {
-        get: function (target: any, prop: string) {
-          // Intercept the property access
-          const property = target[prop];
-
-          if (typeof property === "function") {
-            // Wrap the original method with the extra logging behavior
-            return function (...args: any[]) {
-              // Get the sender of the log event by parsing the stack trace
-              const sender = new Error().stack
-                ?.split("\n").at(2)
-                ?.replace(/^.+\((.+?)\).*$/, "$1")
-                .replace("app://obsidian.md/", "");
-
-              logger.log(prop, sender, ...args);
-
-              // Forward the method call to the original `console` method
-              return property.apply(target, args);
-            };
-          }
-
-          // Return non-function properties as is
-          return property;
-        },
-      },
-    );
+    const consoleProxy = new Proxy(window.console, handler);
 
     // Replace the `console` object with the proxy
     window.console = consoleProxy;
 
     // Listen for uncaught exceptions
-    window.addEventListener("error", this.onWindowError);
+    window.addEventListener("error", this.onWindowError.bind(this));
+    window.addEventListener(
+      "unhandledrejection",
+      this.onWindowUnhandledRejection.bind(this),
+    );
+
+    return this;
   }
 
   // Removing the console proxy object and the listener for uncaught errors
   public teardown(): void {
     window.console = WINDOW_CONSOLE;
-    window.removeEventListener("error", this.onWindowError);
-    console.info(this.logger.prefixMsg("Proxy removed"));
+    window.removeEventListener("error", this.onWindowError.bind(this));
+    window.removeEventListener(
+      "unhandledrejection",
+      this.onWindowUnhandledRejection.bind(this),
+    );
+
+    console.info(prefixMsg("Proxy removed"));
+  }
+
+  /**
+   * Adds log events with the specified level and arguments to the log event
+   * queue.
+   *
+   * @param level - The level of the log event.
+   * @param sender - The sender of the log event (e.g., "plugin:whatever").
+   * @param args - The argument(s) to be logged, optional.
+   */
+  public storeEvent(
+    level: string,
+    sender: string | undefined,
+    ...args: any[]
+  ): void {
+    this.queue.push({ timestamp: new Date(), level, sender, args });
   }
 
   /**
@@ -75,10 +96,47 @@ export class ConsoleProxy {
    * @param event - The error event object.
    */
   private onWindowError(event: ErrorEvent): void {
-    const { message, colno, lineno, filename } = event;
-    const logMessage = `${message} (${filename}:${lineno}:${colno})`;
+    const { message, colno, lineno, filename, error } = event;
 
     // Add a `fatal`-level log event to the queue
-    this.logger.log("fatal", logMessage);
+    this.storeEvent(
+      "fatal",
+      `${filename}:${lineno}:${colno}`,
+      error.name,
+      message,
+      error.stack || "(stack trace unavailable)",
+    );
+  }
+
+  /**
+   * Event handler for unhandled exceptions happening in promises. Adds a
+   * "fatal"-level log event to the log event queue.
+   *
+   * @param event - The error event object.
+   */
+  private onWindowUnhandledRejection(event: UnhandledRejectionEvent): void {
+    const error = event.reason;
+
+    if (typeof error === "string") {
+      this.storeEvent(
+        "fatal",
+        "sender:unknown",
+        "Uncaught (in promise)",
+        error,
+      );
+    } else {
+      const { colno, lineno, filename } = error;
+      const sender = (filename && lineno && colno)
+        ? `${filename}:${lineno}:${colno}`
+        : error.stack?.match(/at eval \((.+?)\)/)?.[1] ?? "(undetermined)";
+
+      // Add a `fatal`-level log event to the queue
+      this.storeEvent(
+        "fatal",
+        sender,
+        "Uncaught (in promise)",
+        error.stack || "(stack trace unavailable)",
+      );
+    }
   }
 }